NN refactor

+++ multi-gammon-core/src/main/java/org/akoshterek/backgammon/nn/Neuron.scala
@@ -4,8 +4,8 @@ import java.io.Serializable
 
 trait Neuron extends Serializable {
 
-  protected var _value: Double = _
-  def value: Double = _value
+  protected var _value: Float = _
+  def value: Float = _value
   /**
    * Recomputes the value of this hidden unit, querying it's
    * prior inputs.
